initial commit
This is a first draft of a solution to
https://github.com/jtdaugherty/dbmigrations/issues/28

This new package only contains the MySQL specific parts of the
dbmigrations code, thereby separating the MySQL dependencies from the
main dbmigrations package. The main goal here is to not force people who
want to use dbmigrations with PostgreSQL or any other database backend
to install MySQL components on their systems.